bug update bug delete back

+++ backend/controllers/post.controllers.js
@@ -1,6 +1,7 @@
 const models = require("../models");
 const jwt = require("jsonwebtoken");
 const fs = require("fs");
+const { post } = require("../routes/post.routes");
 
 exports.createPost = (req, res, next) => {
     if (!req.body.title || !req.body.content) {
@@ -53,6 +54,17 @@ exports.getAllPosts = (req, res, next) => {
   };
 
   exports.deletePost = (req, res, next) => {
+    const token = req.headers.authorization.split(" ")[1];
+    const decodedToken = jwt.verify(token, "RANDOM_TOKEN_SECRET");
+    const userId = decodedToken.userId;
+    const isAdmin = decodedToken.isAdmin;
+    const postIdUser = req.params.idUser;
+
+    if (postIdUser != userId || isAdmin === false) {
+      return res.status(400).json({
+          error: "Vous n'avez pas l'autorisation requise pour faire cela",
+      });
+    }
     models.Post.destroy({where: {id: req.params.id}})
       .then(() => {
         res.status(200).json({
@@ -61,6 +73,7 @@ exports.getAllPosts = (req, res, next) => {
       })
       .catch((error) => res.status(400).json({ error }));
   };
+
